HTML Changes
Fixes for sourceforge bugs 4 and 6.
http://sourceforge.net/p/easygui/bugs/

Created a very cheesy index.html at the top, which has no content other
than the index.

Put all the docs in a documentation folder to keep them separate from
the source.

Still more work as well as checking in the source files to make some of
the docs.