Added other docs sections to sphinx build
Added the "easier" docs to sphinx, which include the faq, cookbook, and
the tutorial.  The tutorial isn't complete.  NOTE: This work is all done
in the sphinx main directory. When it is good enough, you need to copy
the files to the 'documentation/sphinx' directory (for now).

Later, I'll adjust it to make Sphinx the main directory as well as,
perhaps, change the build directory _build to their actual destination
(ie 'documentation') if that's possible.