Tutorial images and registration_system demo added
Added remaining images to the tutorial section.  I also cleaned some of
them up as their right side had the remains of another window.

I added a couple of questions that I thought about to the FAQ.

I added the registration_system code to the cookbook page, which seemed
approprate.

Note in the checkin, but I downloaded the existing website from the
sourceforge location.