Modify doc build location, python3.4 fixes
Modified Sphinx make process (make.bat and Makefile) to build html files
in the ../documentation directory instead of the normal _build
directory.  I'm doing this so building docs is a one-step process and
the build files immediately show up in 'documentation.

Modified a bit to run in python3.4.2 also.  Changed a couple of prints
to methods and redefined basestring.