fixed so import and sphinx docs generate
changed imports to be relative to the directory of the particular
file....hopefully I got it correct.  The demo runs inside of pycharm
both python 2.7 and 3.4.  It also now runs in 2.7 from the command line.

My intent for this checkin is to get master pretty much working fine and
then start doing the development on the develop branch, with incremental
merges back to master.