import fixes to perfect the package
Made many changes to the package.  i really wanted to be able to run
each python file by itself, but after lots of reading and lots of
hacking, it isn't practically possible.

Its much better to write another short script that imports and runs.

For all these modules (except easygui.py), all imports are relative.
This is the "correct" way to do it.

Everything is fairly decent.  There is still the issue with the main gui
not going away on exit/cancel.  but that is unrelated.