Docs and code cleanups
Removed 'docs' directory.  The goal is to have things built using
readthedocs.org

In the code made suggestions from PIP8 and pycharm including changing
arguments in the demo functions.

Changed integerbox default to be of None instead of empty string

Removed some, but not all places where imports are done at the top with
*.  Will adjust more later.

Changed classname in alt_textbox to be camelcase per PIP8.

Added an explicit __all__ to __init__.py  This isn't needed to run
easygui, but for sphinx to autodoc properly each thing we wish to expose
needs to be in __all__