Fix load_tk_image to specify its master (sometimes needed)
Checked in a change to utils.py that supports a tk_master parameter in
load_tk_images.  This is needed when you might have multiple master Tk()
instances.  Not needed for the stock dialogs, but for the
persistent_button_box, it is needed.  Also checked in that prototype
code.  It is not quite done, but when it is, we will always return a
complex object instead of just a string.  This acts like a string in
most ways, but also complex everything related to the GUI, so it also
has a .run() event that will call the GUI up again.