Remove Debug Print
The Debug Print in the "preselect_choice" Function constantly prints an zero array into the console when called "[0]" ...this is slightly unfortunate when trying to keep a neat and tidy console output. A removal would be appreciated. I'd just change it myself in the site-packages but whoever wants to use my code and pip-install your library with it will have the same problem again. Thanks in advance! :)